chore: move MultipleOperatorOut schema to v2
