+++ bc_obps/registration/schema/v2/operation.py
@@ -2,10 +2,9 @@ from uuid import UUID
 from registration.models.bc_greenhouse_gas_id import BcGreenhouseGasId
 from registration.models.bc_obps_regulated_operation import BcObpsRegulatedOperation
 from typing import List, Optional, Literal
-from registration.schema.v1.multiple_operator import MultipleOperatorOut
 from registration.models.contact import Contact
 from registration.schema.v1.operator import OperatorForOperationOut
-from registration.schema.v2.multiple_operator import MultipleOperatorIn
+from registration.schema.v2.multiple_operator import MultipleOperatorIn, MultipleOperatorOut
 from ninja import Field, ModelSchema, Schema
 from registration.models import MultipleOperator, Operation
 from registration.models.opted_in_operation_detail import OptedInOperationDetail
